phishing_kit/image_hashing_storage/requirements.txt
scipy
tldextract
opencv-python-headless
pandas
numpy
tqdm
Pillow==8.4.0
pathlib
fvcore
pycocotools
scikit-learn
lxml
gdown
editdistance
lxml
cryptography==38.0.4
httpcore==0.15.0
h11
h2
hyperframe
selenium==4.0.0
selenium-wire
boto3
imagehash
numpy
onnxruntime
requests
torch